backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
httpx>=0.26.0
python-dotenv>=1.0.0
tenacity>=8.2.0
playwright>=1.41.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.20.0
PyExecJS>=1.5.1
